Update readme file
Replace address from project codeship with gitlab to a new address
project codeship with github.
Replace address from repository gitlab to
repository github

As the intention is to maintain this project only in github. I created a
repository in github and move all code from gitlab to github. It was
necessary to create a new project in codeship and because of that
the address from older codeshipt project was changed to a new address

As a result of this change, the reference to github and codeship is
correct